+++ lib/temporal.dart
@@ -2,7 +2,7 @@
 export 'src/temporal/date_time/accessors.dart';
 export 'src/temporal/date_time/periodical.dart';
 export 'src/temporal/date_time/truncate.dart';
-export 'src/temporal/duration/accessors.dart';
 export 'src/temporal/duration/convert.dart';
 export 'src/temporal/duration/convert_all.dart';
+export 'src/temporal/interval/duration.dart';
 export 'src/temporal/time_unit.dart';